// src/slate-types.ts
export enum EBlockType {
  Paragraph = 'paragraph',
  ListItem = 'list-item',
  BulletedList = 'bulleted-list',
  NumberedList = 'numbered-list',
  Table = 'table',
  TableHeader = 'table-header',
  TableBody = 'table-body',
  TableFooter = 'table-footer',
  TableRow = 'table-row',
  TableHeaderCell = 'table-header-cell',
  TableCell = 'table-cell',
  BlockQuote = 'block-quote',
  Code = 'code',
  HeadingOne = 'heading-one',
  HeadingTwo = 'heading-two',
  HeadingThree = 'heading-three',
  HeadingFour = 'heading-four',
  HeadingFive = 'heading-five',
  HeadingSix = 'heading-six',
  HorizontalRule = 'horizontal-rule'
}

export enum EInlineType {
  CodeLine = 'code-line',
  Image = 'image',
  Input = 'input'
}

export enum EMarkType {
  Bold = 'bold',
  Italic = 'italic',
  Underline = 'underline',
  Strikethrough = 'strikethrough',
  Link = 'link'
}